requirements.txt
asgiref==3.7.2
cfgv==3.4.0
distlib==0.3.7
Django==4.2.7
django-stubs==4.2.6
django-stubs-ext==4.2.5
filelock==3.13.1
identify==2.5.31
mypy==1.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
platformdirs==3.11.0
pre-commit==3.5.0
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
typing_extensions==4.8.0
virtualenv==20.24.6
djangorestframework==3.15.1
